Use PoolOffset everywhere a pointer was used
The code is modified such that evey place where a pointer to a certain
object is used, an offset/index into a suitable memory pool is used
instead.

One advantege to this is that simple random-access arrays can be used to
map corresponding entities (for example, mapping a regex expr to its NFA).

One disadvantage is that pointer types might help in readability and type
checking. Enhancing readability is easy since we can typedef PoolOffset
to different aliases to reflect the pool we are accessing through any
variable.